Fix Bugs contract controller global controller e importation controller

+++ app/Http/Controllers/GlobalChargesLclController.php
@@ -269,7 +269,10 @@ class GlobalChargesLclController extends Controller
     public function destroyGlobalCharges($id)
     {
         $global = GlobalChargeLcl::find($id);
-        $global->delete();
+        if(!empty($global) == 0){
+            $global->delete();
+        }
+        
     }
 
     public function destroyArr(Request $request)
